Updated the DBN.
Now it also contains as nodes the actions for the traffic light per
lane.

The traffic of each lane at time t+1 depends on the traffic of the lane
at time t and at the action happened at time t. The action at t+1
depends on the action at t and at the traffic of the lane at t+1
(synchronic arc).