Removed separate action for each lane.
The state at t+1 for the lanes depend on the whole set of traffic light
conditions as well as the condition at t for that lane. 

Also removed the synchronic arc for the action at t+1. Now the action at
t+1 depends on the action at t and the condition of all the lanes at
time t. 

(Each action is considered the set of traffic light configurations for
an intersection)